+++ Sources/ScreamURITemplate/Internal/Scanner.swift
@@ -1,4 +1,4 @@
-//   Copyright 2018-2024 Alex Deem
+//   Copyright 2018-2025 Alex Deem
 //
 //   Licensed under the Apache License, Version 2.0 (the "License");
 //   you may not use this file except in compliance with the License.
@@ -33,7 +33,7 @@ struct Scanner {
         return currentIndex >= unicodeScalars.endIndex
     }
 
-    mutating func scanComponent() throws -> Component {
+    mutating func scanComponent() throws(URITemplate.Error) -> Component {
         let nextScalar = unicodeScalars[currentIndex]
 
         switch nextScalar {
@@ -44,11 +44,11 @@ struct Scanner {
         case literalCharacterSet:
             return try scanLiteralComponent()
         default:
-            throw URITemplate.Error.malformedTemplate(position: currentIndex, reason: "Unexpected character")
+            throw URITemplate.Error(type: .malformedTemplate, position: currentIndex, reason: "Unexpected character")
         }
     }
 
-    private mutating func scanExpressionComponent() throws -> Component {
+    private mutating func scanExpressionComponent() throws(URITemplate.Error) -> Component {
         assert(unicodeScalars[currentIndex] == "{")
         let expressionStartIndex = currentIndex
         currentIndex = unicodeScalars.index(after: currentIndex)
@@ -59,11 +59,11 @@ struct Scanner {
         return ExpressionComponent(expressionOperator: expressionOperator, variableList: variableList, templatePosition: expressionStartIndex)
     }
 
-    private mutating func scanExpressionOperator() throws -> ExpressionOperator {
+    private mutating func scanExpressionOperator() throws(URITemplate.Error) -> ExpressionOperator {
         let expressionOperator: ExpressionOperator
         if expressionOperatorCharacterSet.contains(unicodeScalars[currentIndex]) {
             guard let `operator` = ExpressionOperator(rawValue: unicodeScalars[currentIndex]) else {
-                throw URITemplate.Error.malformedTemplate(position: currentIndex, reason: "Unsupported Operator")
+                throw URITemplate.Error(type: .malformedTemplate, position: currentIndex, reason: "Unsupported Operator")
             }
             expressionOperator = `operator`
             currentIndex = unicodeScalars.index(after: currentIndex)
@@ -73,7 +73,7 @@ struct Scanner {
         return expressionOperator
     }
 
-    private mutating func scanVariableList() throws -> [VariableSpec] {
+    private mutating func scanVariableList() throws(URITemplate.Error) -> [VariableSpec] {
         var variableList: [VariableSpec] = []
 
         var complete = false
@@ -81,13 +81,13 @@ struct Scanner {
             let variableName = try scanVariableName()
 
             if currentIndex == unicodeScalars.endIndex {
-                throw URITemplate.Error.malformedTemplate(position: currentIndex, reason: "Unterminated Expression")
+                throw URITemplate.Error(type: .malformedTemplate, position: currentIndex, reason: "Unterminated Expression")
             }
 
             let modifier = try scanVariableModifier()
 
             if currentIndex == unicodeScalars.endIndex {
-                throw URITemplate.Error.malformedTemplate(position: currentIndex, reason: "Unterminated Expression")
+                throw URITemplate.Error(type: .malformedTemplate, position: currentIndex, reason: "Unterminated Expression")
             }
 
             variableList.append(VariableSpec(name: variableName, modifier: modifier))
@@ -99,20 +99,20 @@ struct Scanner {
                 currentIndex = unicodeScalars.index(after: currentIndex)
                 complete = true
             default:
-                throw URITemplate.Error.malformedTemplate(position: currentIndex, reason: "Unexpected Character in Expression")
+                throw URITemplate.Error(type: .malformedTemplate, position: currentIndex, reason: "Unexpected Character in Expression")
             }
         }
 
         return variableList
     }
 
-    private mutating func scanVariableName() throws -> Substring {
+    private mutating func scanVariableName() throws(URITemplate.Error) -> Substring {
         let endIndex = scanUpTo(characterSet: invertedVarnameCharacterSet)
         let variableName = string[currentIndex..<endIndex]
         if variableName.isEmpty {
-            throw URITemplate.Error.malformedTemplate(position: currentIndex, reason: "Empty Variable Name")
+            throw URITemplate.Error(type: .malformedTemplate, position: currentIndex, reason: "Empty Variable Name")
         } else if variableName.starts(with: ".") {
-            throw URITemplate.Error.malformedTemplate(position: currentIndex, reason: "Variable Name Cannot Begin With '.'")
+            throw URITemplate.Error(type: .malformedTemplate, position: currentIndex, reason: "Variable Name Cannot Begin With '.'")
         }
         var remainingVariableName = variableName
         while let index = remainingVariableName.firstIndex(of: "%") {
@@ -120,7 +120,7 @@ struct Scanner {
             let thirdIndex = remainingVariableName.index(after: secondIndex)
             if !hexCharacterSet.contains(unicodeScalars[secondIndex]) ||
                 !hexCharacterSet.contains(unicodeScalars[thirdIndex]) {
-                throw URITemplate.Error.malformedTemplate(position: currentIndex, reason: "% must be percent-encoded in variable name")
+                throw URITemplate.Error(type: .malformedTemplate, position: currentIndex, reason: "% must be percent-encoded in variable name")
             }
             let nextIndex = remainingVariableName.index(after: thirdIndex)
             remainingVariableName = remainingVariableName[nextIndex...]
@@ -129,7 +129,7 @@ struct Scanner {
         return variableName
     }
 
-    private mutating func scanVariableModifier() throws -> VariableSpec.Modifier {
+    private mutating func scanVariableModifier() throws(URITemplate.Error) -> VariableSpec.Modifier {
         switch unicodeScalars[currentIndex] {
         case "*":
             currentIndex = unicodeScalars.index(after: currentIndex)
@@ -139,16 +139,16 @@ struct Scanner {
             let endIndex = scanUpTo(characterSet: invertedDecimalDigitsCharacterSet)
             let lengthString = string[currentIndex..<endIndex]
             if lengthString.isEmpty {
-                throw URITemplate.Error.malformedTemplate(position: currentIndex, reason: "Prefix length not specified")
+                throw URITemplate.Error(type: .malformedTemplate, position: currentIndex, reason: "Prefix length not specified")
             }
             if lengthString.first == "0" {
-                throw URITemplate.Error.malformedTemplate(position: currentIndex, reason: "Prefix length cannot begin with 0")
+                throw URITemplate.Error(type: .malformedTemplate, position: currentIndex, reason: "Prefix length cannot begin with 0")
             }
             if lengthString.count > 4 {
-                throw URITemplate.Error.malformedTemplate(position: currentIndex, reason: "Prefix modifier length too large")
+                throw URITemplate.Error(type: .malformedTemplate, position: currentIndex, reason: "Prefix modifier length too large")
             }
             guard let length = Int(lengthString) else {
-                throw URITemplate.Error.malformedTemplate(position: currentIndex, reason: "Cannot parse prefix modifier length")
+                throw URITemplate.Error(type: .malformedTemplate, position: currentIndex, reason: "Cannot parse prefix modifier length")
             }
             currentIndex = endIndex
             return .prefix(length: length)
@@ -157,7 +157,7 @@ struct Scanner {
         }
     }
 
-    private mutating func scanLiteralComponent() throws -> Component {
+    private mutating func scanLiteralComponent() throws(URITemplate.Error) -> Component {
         assert(literalCharacterSet.contains(unicodeScalars[currentIndex]))
 
         let startIndex = currentIndex
@@ -166,7 +166,7 @@ struct Scanner {
         return LiteralComponent(string[startIndex..<endIndex])
     }
 
-    private mutating func scanPercentEncodingComponent() throws -> Component {
+    private mutating func scanPercentEncodingComponent() throws(URITemplate.Error) -> Component {
         assert(unicodeScalars[currentIndex] == "%")
 
         let startIndex = currentIndex
@@ -175,7 +175,7 @@ struct Scanner {
 
         if !hexCharacterSet.contains(unicodeScalars[secondIndex]) ||
             !hexCharacterSet.contains(unicodeScalars[thirdIndex]) {
-            throw URITemplate.Error.malformedTemplate(position: currentIndex, reason: "% must be percent-encoded in literal")
+            throw URITemplate.Error(type: .malformedTemplate, position: currentIndex, reason: "% must be percent-encoded in literal")
         }
 
         currentIndex = unicodeScalars.index(after: thirdIndex)