+++ Sources/ScreamURITemplate/URITemplate.swift
@@ -1,4 +1,4 @@
-//   Copyright 2018-2024 Alex Deem
+//   Copyright 2018-2025 Alex Deem
 //
 //   Licensed under the Apache License, Version 2.0 (the "License");
 //   you may not use this file except in compliance with the License.
@@ -17,11 +17,20 @@ import Foundation
 /// An [RFC6570](https://tools.ietf.org/html/rfc6570) URI Template
 public struct URITemplate {
     /// An error that may be thrown when parsing or processing a template
-    public enum Error: Swift.Error {
-        /// Represents an error parsing a string into a URI Template
-        case malformedTemplate(position: String.Index, reason: String)
-        /// Represents an error processing a template
-        case expansionFailure(position: String.Index, reason: String)
+    public struct Error: Swift.Error {
+        public enum ErrorType: Sendable {
+            /// Represents an error parsing a string into a URI Template
+            case malformedTemplate
+            /// Represents an error processing a template
+            case expansionFailure
+        }
+
+        /// The type of the error
+        public let type: ErrorType
+        /// The position in the template that the error occurred
+        public let position: String.Index
+        /// The reason for the error
+        public let reason: String
     }
 
     private let string: String
@@ -30,8 +39,8 @@ public struct URITemplate {
     /// Initializes a URITemplate from a string
     /// - Parameter string: the string representation of the URI Template
     ///
-    /// - Throws: `URITemplate.Error.malformedTemplate` if the string is not a valid URI Template
-    public init(string: String) throws {
+    /// - Throws: `URITemplate.Error` with `type = .malformedTemplate` if the string is not a valid URI Template
+    public init(string: String) throws(URITemplate.Error) {
         var components: [Component] = []
         var scanner = Scanner(string: string)
         while !scanner.isComplete {
@@ -46,8 +55,8 @@ public struct URITemplate {
     ///
     /// - Returns: The result of processing the template
     ///
-    /// - Throws: `URITemplate.Error.expansionFailure` if an error occurs processing the template
-    public func process(variables: TypedVariableProvider) throws -> String {
+    /// - Throws: `URITemplate.Error` with `type = .expansionFailure` if an error occurs processing the template
+    public func process(variables: TypedVariableProvider) throws(URITemplate.Error) -> String {
         var result = ""
         for component in components {
             result += try component.expand(variables: variables)
@@ -63,8 +72,8 @@ public struct URITemplate {
     ///
     /// - Returns: The result of processing the template
     ///
-    /// - Throws: `URITemplate.Error.expansionFailure` if an error occurs processing the template
-    public func process(variables: VariableProvider) throws -> String {
+    /// - Throws: `URITemplate.Error` with `type = .expansionFailure` if an error occurs processing the template
+    public func process(variables: VariableProvider) throws(URITemplate.Error) -> String {
         struct TypedVariableProviderWrapper: TypedVariableProvider {
             let variables: VariableProvider
 
@@ -84,8 +93,8 @@ public struct URITemplate {
     ///
     /// - Returns: The result of processing the template
     ///
-    /// - Throws: `URITemplate.Error.expansionFailure` if an error occurs processing the template
-    public func process(variables: [String: String]) throws -> String {
+    /// - Throws: `URITemplate.Error` with `type = .expansionFailure` if an error occurs processing the template
+    public func process(variables: [String: String]) throws(URITemplate.Error) -> String {
         return try process(variables: variables as VariableDictionary)
     }