+++ Sources/ScreamURITemplateExample/main.swift
@@ -1,4 +1,4 @@
-//   Copyright 2018-2024 Alex Deem
+//   Copyright 2018-2025 Alex Deem
 //
 //   Licensed under the Apache License, Version 2.0 (the "License");
 //   you may not use this file except in compliance with the License.
@@ -14,6 +14,7 @@
 
 import Foundation
 import ScreamURITemplate
+import ScreamURITemplateMacros
 
 let template = try URITemplate(string: "https://api.github.com/repos/{owner}/{repo}/collaborators/{username}")
 let variables = [
@@ -27,3 +28,24 @@ let urlString = try template.process(variables: variables)
 let url = URL(string: urlString)!
 print("Expanding \(template)\n     with \(variables):\n")
 print(url.absoluteString)
+
+let macroExpansion = #URITemplate("https://api.github.com/repos/{owner}/{repo}/collaborators/{username}")
+print(macroExpansion)
+
+let urlExpansion = #URLByExpandingURITemplate("https://api.github.com/repos/{owner}/{repo}/collaborators/{username}",
+                                              with: ["owner": "SwiftScream", "repo": "URITemplate", "username": "alexdeem"])
+print(urlExpansion)
+
+@VariableProvider
+struct GitHubRepoCollaborator {
+    let owner: String
+    let repo: String
+    let username: String
+}
+
+let expansion = try macroExpansion.process(variables: GitHubRepoCollaborator(owner: "SwiftScream", repo: "URITemplate", username: "alexdeem"))
+print(expansion)
+
+let typedTemplate = TypedURITemplate<GitHubRepoCollaborator>(macroExpansion)
+let result = try typedTemplate.process(variables: .init(owner: "SwiftScream", repo: "SwiftScream", username: "alexdeem"))
+print(result)